requirements.txt
# Core ML
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# MLflow
mlflow>=2.10.0

# BERT / Sentence Transformers
sentence-transformers>=2.6.0
torch>=2.0.0

# PDF Reading
PyPDF2>=3.0.0
python-docx>=1.0.0

# Web UI
streamlit>=1.32.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
